@@ -14,22 +14,24 @@
   },
   "main": "lib/functions/src/audioapp.js",
   "dependencies": {
-    "@google-cloud/firestore": "4.0.0",
+    "@google-cloud/firestore": "^7.11.0",
     "@google-cloud/storage": "5.18.2",
     "body-parser": "1.19.2",
     "cookie-parser": "1.4.6",
     "express": "4.17.3",
-    "firebase-admin": "9.0.0",
-    "firebase-functions": "3.19.0",
+    "firebase-admin": "^11.0.0",
+    "firebase-functions": "^4.0.0",
     "source-map-support": "0.5.21",
-    "typescript": "4.6.2",
     "useragent": "2.3.0"
   },
   "devDependencies": {
+    "@types/node": "^18.19.80",
     "@types/supertest": "2.0.12",
     "firebase-functions-test": "0.2.0",
     "intern": "4.10.1",
-    "supertest": "6.2.3"
+    "supertest": "6.2.3",
+    "typescript": "^5.8.2",
+    "undici-types": "^7.5.0"
   },
   "private": true
 }
