@@ -1,17 +1,22 @@
 {
   "dependencies": {
+    "@google-cloud/firestore": "^7.11.0",
     "@google-cloud/storage": "5.18.2",
     "@types/jquery": "3.5.14",
     "cookie-parser": "1.4.6",
     "express": "4.17.3",
     "firebase": "^9.6.8",
-    "firebase-admin": "9.0.0",
-    "firebase-functions": "3.19.0",
-    "firebase-tools": "11.29.0",
     "multer": "1.4.4",
-    "typescript": "4.6.2"
+    "firebase-admin": "^11.0.0",
+    "firebase-functions": "^4.0.0",
+    "firebase-tools": "latest"
   },
   "scripts": {
     "build": "tsc"
+  },
+  "devDependencies": {
+    "@types/node": "^18.19.80",
+    "typescript": "^5.8.2",
+    "undici-types": "^7.5.0"
   }
 }
